Fix autotest-deps-glbench gcc-4.7 building problem.
Added explicit converion in initialization list.

TEST=built with gcc-4.7
BUG=None

Reviewed-on: https://gerrit.chromium.org/gerrit/25686
Tested-by: Han Shen <shenhan@chromium.org>
Commit-Ready: Han Shen <shenhan@chromium.org>

+++ autotest/client/deps/glbench/windowmanagertest.cc
@@ -120,9 +120,9 @@ int main(int argc, char* argv[]) {
     kStateExit
   } state = kStateScreenShot1;
   float seconds_delay_for_next_state[] = {
-      FLAGS_screenshot1_sec,
-      FLAGS_screenshot2_sec,
-      FLAGS_cooldown_sec,
+      static_cast<float>(FLAGS_screenshot1_sec),
+      static_cast<float>(FLAGS_screenshot2_sec),
+      static_cast<float>(FLAGS_cooldown_sec),
       0
   };
